requirements.txt
Django>=3.2
django-cors-headers
djangorestframework
celery
redis
